+++ ninja-build_rs/src/nightly.rs
@@ -1,34 +1,144 @@
-//! Extensions to the amazing [autocfg] designed to help with ergonomically and safely handling
-//! experimental features in nightly.
+//! Checking for experimental or stabilised features is prone to subtle errors which create issues
+//! for downstream users and verbose when done properly. This provides extensions to the amazing
+//! [autocfg::AutoCfg] (re-exported via our prelude to make your life easier) to safely identify the
+//! availability of nightly features & handle the future stabilisation process without additional
+//! effort on your part. All while respecting any `allow-feature` whitelists.
+//!
+//! For a list of known features with dedicated probes see [UnstableFeature]
+//!
+//! ## Usage
+//!
+//! ### `build.rs`
+//!
+//! ```rust, should_panic
+//! use ninja_build_rs::prelude::*;
+//!
+//! fn main() -> Result<()> {
+//!     // get a new AutoCfg or provide a valuable error
+//!     // rather than panicing
+//!     let ac = AutoCfg::new()?;
+//!
+//!     // check to see if the downstream crate has defined
+//!     // `unstable.allow-features` in `.cargo/config.toml`.
+//!     // It is mandatory to perform this check and pass the
+//!     // result to any calls to `emit_unstable_feature`
+//!     let allowed_features = cargo_allowed_features()?;
+//!
+//!     // We want to make use of `assert_matches` if it is available
+//!     ac.emit_unstable_feature(assert_matches, &allowed_features);
+//!     //                       ^^^^^^^^^^^^^^ - enum variant to avoid typos
+//!
+//!     Ok(())
+//! }
+//! ```
+//!
+//! ### `lib.rs` / `main.rs`
+//!
+//! ```rust, ignore
+//! // only enable unstable feature if it is available and has not yet been stabilised
+//! #![cfg_attr(unstable_assert_matches, feature(assert_matches))]
+//!
+//! #[cfg(test)]
+//! // Do these tests if `assert_matches` is available
+//! #[cfg(has_assert_matches)]
+//! mod tests {
+//!     // `assert_matches` was moved in early 2026 before stabilisation
+//!     #[cfg(assert_matches_location = "root")]
+//!     use std::assert_matches;
+//!
+//!     // in earlier nightly compilers `assert_matches` was in a separate module
+//!     #[cfg(assert_matches_location = "module")]
+//!     use std::assert_matches::assert_matches;
+//!
+//!     #[test]
+//!     fn has() {
+//!         assert_matches!(Some(5), Some(n) if n == 5);
+//!     }
+//! }
+//!
+//! #[cfg(test)]
+//! // Do these tests if `assert_matches` is not available
+//! #[cfg(not(has_assert_matches))]
+//! mod tests {
+//!     #[test]
+//!     fn has_not() {
+//!         assert_eq!(Some(5), Some(5));
+//!     }
+//! }
+//! ```
+//!
+//! ## Note to downstream crates
+//!
+//! If you (transiently) depend on a crate which uses `ninja-build_rs` and have implemented a
+//! whitelist of `allowed-features`.
+//!
+//! Due to limitations in the information provided by cargo:
+//!
+//! - This will obtain config.toml files based upon `OUT_DIR`. If this is not under the project
+//!   root, you can override by providing an alternative path via the environment variable
+//!   `NINJA_CARGO_CONFIG_DIR`. See cargo's documentation on config file hierarchical structure
+//!   for more details.
+//! - This will not respect additional entries passed at the command line via
+//!   `cargo --config unstable.allow-features=[...]`
 
 use std::{
+    fmt::Debug,
     path::Path,
     process::{Command, Output},
 };
 
-use autocfg::AutoCfg;
+pub use autocfg::AutoCfg;
+use derive_more::Display;
 
 use crate::{BuildError, Result, get_var};
-use probes::{has, make_probe};
+use probes::{has, make_probe, unstable};
 
-/// Known features with `unstable_...` & `has_...`
+/// Known features with `unstable_...` & `has_...`.
+///
+/// If the feature you want is not in this list you can use `Other` to get `unstable_...`
+/// but please also raise a PR (or open an issue) to add a custom probe for `has_...`.
 #[allow(non_camel_case_types, reason = "shadowing feature naming")]
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+#[derive(Debug, Clone, PartialEq, Eq, Display)]
 pub enum UnstableFeature {
-    /// Also offers `assert_matches_location=root/module` to identify whether to
-    /// `use std::assert_matches`(root) or `use std::assert_matches::assert_matches` (module)
+    /// ### Provides cfg flags:
+    /// - `#![cfg_attr(unstable_assert_matches, feature(assert_matches))]`
+    /// - `#[cfg(has_assert_matches)]`
+    /// - ```rust, ignore
+    ///   #[cfg(assert_matches_location = "root")]
+    ///   use std::assert_matches;
+    ///   ```
+    /// - ```rust, ignore
+    ///   #[cfg(assert_matches_location = "module")]
+    ///   use std::assert_matches::assert_matches;
+    ///   ```
     assert_matches,
+    /// ### Provides cfg flags:
+    /// - `#![cfg_attr(unstable_iterator_try_collect, feature(iterator_try_collect))]`
+    /// - `#[cfg(has_iterator_try_collect)]`
     iterator_try_collect,
+    /// ### Provides cfg flags:
+    /// - `#![cfg_attr(unstable_never_type, feature(never_type))]`
+    /// - `#[cfg(has_never_type)]`
     never_type,
+    /// ### Provides cfg flags:
+    /// - `#![cfg_attr(unstable_proc_macro_diagnostic, feature(proc_macro_diagnostic))]`
+    /// - `#[cfg(has_proc_macro_diagnostic)]`
     proc_macro_diagnostic,
+    /// ### Provides cfg flags:
+    /// - `#![cfg_attr(unstable_try_trait_v2, feature(try_trait_v2))]`
+    /// - `#[cfg(has_try_trait_v2)]`
     try_trait_v2,
+    /// ### Provides cfg flags:
+    /// - `#![cfg_attr(unstable_try_trait_v2_residual, feature(try_trait_v2_residual))]`
+    /// - `#[cfg(has_try_trait_v2_residual)]`
     try_trait_v2_residual,
     /// only provides `unstable_...` - please raise a PR to add a custom probe for `has_...`
-    Other(&'static str),
+    OtherFeature(String),
 }
 
-impl From<&'static str> for UnstableFeature {
-    fn from(feature: &'static str) -> Self {
+impl UnstableFeature {
+    // This is not pub or trait From to avoid risk of typos
+    fn from(feature: &str) -> Self {
         match feature {
             "assert_matches" => Self::assert_matches,
             "iterator_try_collect" => Self::iterator_try_collect,
@@ -36,15 +146,19 @@ impl From<&'static str> for UnstableFeature {
             "proc_macro_diagnostic" => Self::proc_macro_diagnostic,
             "try_trait_v2" => Self::try_trait_v2,
             "try_trait_v2_residual" => Self::try_trait_v2_residual,
-            _ => Self::Other(feature),
+            _ => Self::OtherFeature(feature.to_string()),
         }
     }
 }
 
 mod probes {
-    use autocfg::AutoCfg;
+    use super::{AutoCfg, UnstableFeature};
 
-    pub fn make_probe(feature: &str, allowed: bool, probe: &str) -> String {
+    /// Prefix with:
+    /// - #![allow(stable_features)] (only if allowed)
+    /// - #![feature()] (only if allowed)
+    /// - #![allow(unused)] (always)
+    pub fn make_probe(feature: &UnstableFeature, allowed: bool, probe: &str) -> String {
         let mut _probe = String::with_capacity(256);
         if allowed {
             _probe.push('\n');
@@ -52,7 +166,7 @@ mod probes {
             _probe.push('\n');
 
             _probe.push_str("#![feature(");
-            _probe.push_str(feature);
+            _probe.push_str(&feature.to_string());
             _probe.push_str(")]");
             _probe.push('\n');
         };
@@ -62,7 +176,8 @@ mod probes {
         _probe
     }
 
-    pub fn has(ac: &AutoCfg, feature: &str, allowed: bool, probe: &str) {
+    /// Register `#[cfg(has_feature)]` & set based on the probe
+    pub fn has(ac: &AutoCfg, feature: &UnstableFeature, allowed: bool, probe: &str) {
         let cfg = format!("has_{feature}");
         autocfg::emit_possibility(&cfg);
         let code = make_probe(feature, allowed, probe);
@@ -71,8 +186,27 @@ mod probes {
         }
     }
 
-    pub mod assert_matches {
+    /// Register `#[cfg(has_feature)]` & run a default probe
+    pub fn unstable(ac: &AutoCfg, feature: &UnstableFeature, allowed: bool) {
+        let cfg = format!("unstable_{feature}");
+        autocfg::emit_possibility(&cfg);
 
+        if allowed {
+            let code = format!(
+                r#"
+#![deny(stable_features)]
+#![feature({feature})]
+#![allow(unused)]
+"#
+            );
+
+            if ac.probe_raw(&code).is_ok() {
+                autocfg::emit(&cfg);
+            }
+        }
+    }
+
+    pub mod assert_matches {
         pub const AVAILABLE: &str = r#"
 use std::assert_matches;
 "#;
@@ -83,7 +217,6 @@ fn main() {
     assert_matches!(Some(4), Some(_));
 }
 "#;
-
         // was stabilised in root - so no need to remove feature from this probe
         pub const MODULE: &str = r#"
 #![allow(stable_features)]
@@ -95,6 +228,7 @@ fn main() {
 }
 "#;
     }
+
     pub mod iterator_try_collect {
         // vec! not array: https://internals.rust-lang.org/t/code-compiles-on-playground-but-fails-when-passed-via-stdin-to-rustc/24393
         pub const AVAILABLE: &str = r#"
@@ -103,19 +237,21 @@ fn try_collect() {
 }
 "#;
     }
+
     pub mod never_type {
         pub const AVAILABLE: &str = r#"
 type Bang = !;
 "#;
     }
+
     pub mod proc_macro_diagnostic {
+        /// Special probe as feature only available in proc_macro context
         pub const UNSTABLE: &str = r#"
 #![deny(stable_features)]
 #![feature(proc_macro_diagnostic)]
 #![allow(unused)]
 extern crate proc_macro;
 "#;
-
         pub const AVAILABLE: &str = r#"
 extern crate proc_macro;
 use proc_macro::Diagnostic;      
@@ -144,9 +280,9 @@ pub trait Nightly {
     /// - To be used at top-level crate via `#![cfg_attr(unstable_foo, feature(foo))]`
     ///
     /// ## Cfg-gating `has_...`
-    /// Do **not** rely on `cfg(not(unstable_foo))` to suggest that `feature(foo)` is stable! There are 3
-    /// reasons that `cfg(unstable_foo)` could be `false`:
-    ///   1. The build is using `stable`/`beta`
+    /// Do **not** rely on `cfg(not(unstable_foo))` to suggest that `feature(foo)` is stable! There
+    /// are 3 reasons that `cfg(unstable_foo)` could be `false`:
+    ///   1. The build is using `stable`/`beta` or the feature is not on the `allow-features` whitelist
     ///   2. The feature has been stabilised
     ///   3. The compiler is from before the feature was implemented
     ///
@@ -154,40 +290,23 @@ pub trait Nightly {
     ///
     /// If you need to test that a feature is available in order to cfg-gate your code and it is not
     /// on the list of [known features](UnstableFeature), please raise a PR with a suggested probe.
-    fn emit_unstable_feature(&self, feature: &'static str, allowed_features: &AllowedFeatures);
-}
-
-fn default_unstable_cfg(ac: &AutoCfg, feature: &'static str, allowed: bool) {
-    let cfg = format!("unstable_{feature}");
-    autocfg::emit_possibility(&cfg);
-
-    if allowed {
-        let code = format!(
-            r#"
-#![deny(stable_features)]
-#![feature({feature})]
-#![allow(unused)]
-"#
-        );
-
-        if ac.probe_raw(&code).is_ok() {
-            autocfg::emit(&cfg);
-        }
-    }
+    fn emit_unstable_feature(&self, feature: UnstableFeature, allowed_features: &AllowedFeatures);
 }
 
 impl Nightly for AutoCfg {
-    fn emit_unstable_feature(&self, feature: &'static str, allowed_features: &AllowedFeatures) {
+    fn emit_unstable_feature(&self, feature: UnstableFeature, allowed_features: &AllowedFeatures) {
+        // show in `cargo build -vv`
         dbg!(&feature);
+
         let ac = self;
-        let allowed = allowed_features.includes(feature);
-        match UnstableFeature::from(feature) {
+        let allowed = allowed_features.includes(&feature);
+        match feature {
             UnstableFeature::assert_matches => {
-                default_unstable_cfg(self, feature, allowed);
-                has(ac, feature, allowed, probes::assert_matches::AVAILABLE);
+                unstable(self, &feature, allowed);
+                has(ac, &feature, allowed, probes::assert_matches::AVAILABLE);
                 autocfg::emit_possibility("assert_matches_location, values(\"root\", \"module\")");
                 if self
-                    .probe_raw(&make_probe(feature, allowed, probes::assert_matches::ROOT))
+                    .probe_raw(&make_probe(&feature, allowed, probes::assert_matches::ROOT))
                     .is_ok()
                 {
                     autocfg::emit("assert_matches_location=\"root\"")
@@ -197,17 +316,17 @@ impl Nightly for AutoCfg {
                 }
             }
             UnstableFeature::iterator_try_collect => {
-                default_unstable_cfg(self, feature, allowed);
+                unstable(self, &feature, allowed);
                 has(
                     ac,
-                    feature,
+                    &feature,
                     allowed,
                     probes::iterator_try_collect::AVAILABLE,
                 );
             }
             UnstableFeature::never_type => {
-                default_unstable_cfg(self, feature, allowed);
-                has(ac, feature, allowed, probes::never_type::AVAILABLE);
+                unstable(self, &feature, allowed);
+                has(ac, &feature, allowed, probes::never_type::AVAILABLE);
             }
             UnstableFeature::proc_macro_diagnostic => {
                 autocfg::emit_possibility("unstable_proc_macro_diagnostic");
@@ -220,29 +339,31 @@ impl Nightly for AutoCfg {
                 }
                 has(
                     ac,
-                    feature,
+                    &feature,
                     allowed,
                     probes::proc_macro_diagnostic::AVAILABLE,
                 );
             }
             UnstableFeature::try_trait_v2 => {
-                default_unstable_cfg(self, feature, allowed);
-                has(ac, feature, allowed, probes::try_trait_v2::AVAILABLE);
+                unstable(self, &feature, allowed);
+                has(ac, &feature, allowed, probes::try_trait_v2::AVAILABLE);
             }
             UnstableFeature::try_trait_v2_residual => {
-                default_unstable_cfg(self, feature, allowed);
+                unstable(self, &feature, allowed);
                 has(
                     ac,
-                    feature,
+                    &feature,
                     allowed,
                     probes::try_trait_v2_residual::AVAILABLE,
                 );
             }
-            UnstableFeature::Other(feature) => default_unstable_cfg(self, feature, allowed),
+            UnstableFeature::OtherFeature(_) => unstable(self, &feature, allowed),
         }
     }
 }
 
+/// Check whether cargo will accept unstable flags. You probably never need to run this
+/// yourself and should prefer to simply call [`cargo_allowed_features`].
 pub fn cargo_unstable() -> Result<bool> {
     Ok(Command::new(get_var("CARGO")?)
         .args([
@@ -360,7 +481,7 @@ fn _cargo_allowed_features<P: AsRef<Path>>(current_dir: Option<P>) -> Result<All
                 .split(",")
                 .map(str::trim)
                 .filter(|feature| !added_unstable_options || *feature != "unstable-options")
-                .map(ToString::to_string)
+                .map(UnstableFeature::from)
                 .collect();
             if features.is_empty() {
                 AllowedFeatures(_AllowedFeatures::None)
@@ -385,7 +506,7 @@ pub struct AllowedFeatures(_AllowedFeatures);
 impl AllowedFeatures {
     /// Not public as this doesn't consider any restrictions made via `RUSTFLAGS`, those
     /// features will be disabled for all calls to rustc when running probes.
-    fn includes(&self, feature: &str) -> bool {
+    fn includes(&self, feature: &UnstableFeature) -> bool {
         match &self.0 {
             _AllowedFeatures::None => false,
             _AllowedFeatures::All => true,
@@ -395,10 +516,11 @@ impl AllowedFeatures {
 }
 
 #[derive(Debug, Clone, PartialEq, Eq)]
+/// private to make it impossible to manually construct an [AllowedFeatures] from outside this crate
 enum _AllowedFeatures {
     None,
     All,
-    Some(Vec<String>),
+    Some(Vec<UnstableFeature>),
 }
 
 #[cfg(test)]
@@ -411,6 +533,7 @@ mod tests {
 
     use tempfile::TempDir;
 
+    use super::UnstableFeature::*;
     use super::*;
 
     #[test]
@@ -419,10 +542,10 @@ mod tests {
         let allowed = _cargo_allowed_features(Some(&tmp));
         if cargo_unstable().expect("cargo_unstable") {
             assert_matches!(allowed, Ok(AllowedFeatures(_AllowedFeatures::All)));
-            assert!(allowed.unwrap().includes("try_trait_v2"));
+            assert!(allowed.unwrap().includes(&try_trait_v2));
         } else {
             assert_matches!(allowed, Ok(AllowedFeatures(_AllowedFeatures::None)));
-            assert!(!allowed.unwrap().includes("try_trait_v2"));
+            assert!(!allowed.unwrap().includes(&try_trait_v2));
         }
     }
 
@@ -445,14 +568,14 @@ mod tests {
             assert_matches!(
                 allowed,
                 AllowedFeatures(_AllowedFeatures::Some(ref features))
-                if features == &vec!["try_trait_v2", "unstable-options"]
+                if features == &vec![try_trait_v2, OtherFeature("unstable-options".to_string())]
             );
-            assert!(allowed.includes("try_trait_v2"));
-            assert!(allowed.includes("unstable-options"));
+            assert!(allowed.includes(&try_trait_v2));
+            assert!(allowed.includes(&OtherFeature("unstable-options".to_string())));
         } else {
             assert_matches!(allowed, AllowedFeatures(_AllowedFeatures::None));
-            assert!(!allowed.includes("try_trait_v2"));
-            assert!(!allowed.includes("unstable-options"));
+            assert!(!allowed.includes(&try_trait_v2));
+            assert!(!allowed.includes(&OtherFeature("unstable-options".to_string())));
         }
     }
 
@@ -471,12 +594,12 @@ mod tests {
             assert_matches!(
                 allowed,
                 AllowedFeatures(_AllowedFeatures::Some(ref features))
-                if features == &vec!["try_trait_v2"]
+                if features == &vec![try_trait_v2]
             );
-            assert!(allowed.includes("try_trait_v2"));
+            assert!(allowed.includes(&try_trait_v2));
         } else {
             assert_matches!(allowed, AllowedFeatures(_AllowedFeatures::None));
-            assert!(!allowed.includes("try_trait_v2"));
+            assert!(!allowed.includes(&try_trait_v2));
         }
     }
 
@@ -507,6 +630,15 @@ use std::assert_matches;
 use std::assert_matches;
 "#;
 
-        assert_eq!(probes::make_probe("assert_matches", true, probe), expected);
+        assert_eq!(probes::make_probe(&assert_matches, true, probe), expected);
+    }
+
+    #[test]
+    fn unstable_feature_display() {
+        assert_eq!(
+            "foo",
+            format!("{}", UnstableFeature::OtherFeature("foo".to_string()))
+        );
+        assert_eq!("try_trait_v2", format!("{}", UnstableFeature::try_trait_v2))
     }
 }